[ALPS05652867] mml: fix tile index
- Fix tile index type to u32
  while number of tiles may exceed 256 in racing mode
- Fixes of coding convention in core

MTK-Commit-Id: af4c1ded994061b18820eb8d44f02bab9bd7a7f2

CR-Id: ALPS05652867
Feature: [Module]MML driver

+++ drivers/gpu/drm/mediatek/mml/mtk-mml-color.c
@@ -261,7 +261,7 @@ static s32 color_config_frame(struct mml_comp *comp, struct mml_task *task,
 }
 
 static s32 color_config_tile(struct mml_comp *comp, struct mml_task *task,
-			     struct mml_comp_config *ccfg, u8 idx)
+			     struct mml_comp_config *ccfg, u32 idx)
 {
 	struct mml_frame_config *cfg = task->config;
 	struct cmdq_pkt *pkt = task->pkts[ccfg->pipe];
