+++ drivers/gpu/drm/mediatek/mml/mtk-mml-fg.c
@@ -130,7 +130,7 @@ static s32 fg_config_frame(struct mml_comp *comp, struct mml_task *task,
 }
 
 static s32 fg_config_tile(struct mml_comp *comp, struct mml_task *task,
-			  struct mml_comp_config *ccfg, u8 idx)
+			  struct mml_comp_config *ccfg, u32 idx)
 {
 	struct mml_frame_config *cfg = task->config;
 	struct cmdq_pkt *pkt = task->pkts[ccfg->pipe];
