+++ drivers/gpu/drm/mediatek/mml/mtk-mml-rdma.c
@@ -798,7 +798,7 @@ static s32 rdma_config_frame(struct mml_comp *comp, struct mml_task *task,
 }
 
 static s32 rdma_config_tile(struct mml_comp *comp, struct mml_task *task,
-			    struct mml_comp_config *ccfg, u8 idx)
+			    struct mml_comp_config *ccfg, u32 idx)
 {
 	struct mml_frame_config *cfg = task->config;
 	struct rdma_frame_data *rdma_frm = rdma_frm_data(ccfg);
